requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
sqlmodel==0.0.16
asyncpg==0.29.0
psycopg2-binary==2.9.9
argon2-cffi==23.1.0

